Fixed "Overload resolution failed" Error
1. added `dtype=np.uint8` at `output_image = np.full((OUTPUT_IMAGE_DIMENSIONS.y, OUTPUT_IMAGE_DIMENSIONS.x, 3), 255)` in function `seed_to_pattern` in `seed_to_pattern.py`
2. changed `SCALAR = 30` to `SCALAR = 8` in `main.py`
3. rewrite `draw_tile` in `graphics_cv2.py`

+++ code/pattern_generation/seed_to_pattern.py
@@ -51,7 +51,7 @@ def seed_to_pattern(seed, output_file_name="output.png"):
     next_generation()
     
     while True:
-        output_image = np.full((OUTPUT_IMAGE_DIMENSIONS.y, OUTPUT_IMAGE_DIMENSIONS.x, 3), 255)
+        output_image = np.full((OUTPUT_IMAGE_DIMENSIONS.y, OUTPUT_IMAGE_DIMENSIONS.x, 3), 255, dtype=np.uint8)
         for tile in vertices_to_draw:
             output_image = draw_tile(tile, output_image, offset_coord=offset_coordinate)
